Script is working and can run case

+++ Modules/AddtionalFunctions.py
@@ -1,16 +1,18 @@
 import os
 
-def changeVariablesFunV2(distVar, sourVar, nameFile=' '):
+def changeVariablesFunV2(distVar, sourVar, nameFile=''):
     """Function to find and replace required text part at given file
     distVar depicts finding variables
     sourVar depicts replacing variables
     nameFile is the name of file where the procedure will be done"""
-    with open (nameFile, 'r') as f:
-        oldData = f.read()
-    newData = oldData.replace(str(distVar), str(sourVar))
 
-    with open(nameFile, 'w') as f:
-        f.write(newData)
+
+    if os.path.isfile(nameFile):
+        with open (nameFile, 'r') as f:
+            oldData = f.read()
+        newData = oldData.replace(str(distVar), str(sourVar))
+        with open(nameFile, 'w') as f:
+            f.write(newData)
 
 def changeVariablesFun(distVar, sourVar, nameFile=' '):
     """The fucntion is devoted to change  gotten text part
